update fmt, trying to get textfield inputs working for non english

+++ engine/buffer.h
@@ -43,7 +43,7 @@ struct BufferElem {
             case BufferType::Mat4:
                 return GL_FLOAT;
             default:
-                log_error("Missing type conversion for {}", type);
+                log_error("Missing type conversion for {}", (int)type);
                 return GL_FLOAT;
         }
     }
@@ -65,7 +65,7 @@ struct BufferElem {
             case BufferType::Bool:
                 return 1;
             default:
-                log_error("Missing buffer type size for {}", t);
+                log_error("Missing buffer type size for {}", (int)t);
                 return -1;
         }
     }
@@ -86,7 +86,7 @@ struct BufferElem {
             case BufferType::Mat4:
                 return 4;
             default:
-                log_error("Missing count for {}", type);
+                log_error("Missing count for {}", (int)type);
                 return -1;
         }
     }
